# docs/swagger.yaml
basePath: /
definitions:
...
  models.Notification:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      isRead:
        type: boolean
      userID:
        type: string
    type: object
...
host: 109.73.202.151:8080
info:
  contact: {}
  description: ITaM API
  title: LiveCode API
  version: "1.0"
paths:
...
  /auth/api/create_notification:
    post:
      consumes:
      - application/json
      description: Создает новое уведомление
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/models.Notification'
      produces:
      - application/json
      responses:
        "201":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Создать уведомление
      tags:
      - Notifications
...
  /auth/api/delete_notification:
    delete:
      description: Удаляет уведомление по ID
      parameters:
      - description: Notification ID
        in: query
        name: notification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid notification ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Notification not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Удалить уведомление
      tags:
      - Notifications
...
  /auth/api/get_all_notifications:
    get:
      description: Возвращает список всех уведомлений или уведомлений пользователя
        с пагинацией
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of notifications
          schema:
            items:
              $ref: '#/definitions/models.Notification'
            type: array
        "400":
          description: Invalid user ID or pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить все уведомления
      tags:
      - Notifications
...
  /auth/api/get_notification/{notification_id}:
    get:
      description: Возвращает уведомление по его ID
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification data
          schema:
            $ref: '#/definitions/models.Notification'
        "400":
          description: Invalid notification ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Notification not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить уведомление по ID
      tags:
      - Notifications
...
  /auth/api/update_notification:
    patch:
      consumes:
      - application/json
      description: Обновляет существующее уведомление
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/models.Notification'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Обновить уведомление
      tags:
      - Notifications
...
schemes:
- http
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: /api/login
    type: oauth2
swagger: "2.0"